ot_ws/requirements.txt
flask-script
Flask-Testing==0.6.2
gunicorn==19.7.1
flask-cors==3.0.2
dateutils
requests
flask-restplus